pyproject.toml
[tool.poetry]
name = "aqueductcore"
version = "0.0.1"
description = "Aqueduct Core"
authors = ["Aqueduct Team <amir.alavi@riverlane.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.8,< 3.12"
grpcio = "^1.57"
pydantic = {version = "^2.3",extras = ["email"]}
typing-extensions = "^4.7"
protobuf = "^4.24"
uvicorn = {version = "^0.23",extras = ["standard"]}
fastapi = "^0.103"
fastapi-users = {version = "^12.1"}
strawberry-graphql = {version = "^0.205",extras = ["fastapi"]}
pydantic-settings = "^2.0"
sqlalchemy = {version = "^2.0",extras = ["asyncio"]}
fastapi-users-db-sqlalchemy = "^6.0"
asyncpg = "^0.28"
python-multipart = "^0.0"
aiofiles = "^23.2"
streaming-form-data = "^1.13"
gunicorn = "^21.2"
pip-licenses = "^4.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
grpcio-tools = "^1.57"
pylint = "^2.17"
yapf = "^0.40"
pytest-grpc = "^0.8"
mypy = "^1.5"
pytest-cov = "^4.1"
pylint-protobuf = "^0.20"
mypy-protobuf = "^3.5"
pip-licenses = "^4.3"
requests = "^2.31"
pytest-asyncio = "^0.21"
httpx = "^0.25"
mkdocs-material = "^9.4"
mike = "^1.1"
aiosqlite = "^0.19"
alembic = "^1.12"
psycopg2-binary = "^2.9"
mkdocstrings = {version = "^0.24",extras = ["python"]}
nbconvert = "^7.14.2"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
